% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{death_predict}
\alias{death_predict}
\title{Predict Using Trained Model}
\usage{
death_predict(model, newdata, type = c("class", "prob"))
}
\arguments{
\item{model}{A death_model object.}

\item{newdata}{New data (expression matrix or death scores).}

\item{type}{For classification: "class" or "prob".}
}
\value{
Predictions.
}
\description{
Make predictions on new data using a trained death_model.
}
