% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{.calc_class_performance}
\alias{.calc_class_performance}
\title{Internal: Calculate classification performance}
\usage{
.calc_class_performance(actual, predicted, prob_matrix = NULL)
}
\description{
Internal: Calculate classification performance
}
\keyword{internal}
