% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{.calc_reg_performance}
\alias{.calc_reg_performance}
\title{Internal: Calculate regression performance}
\usage{
.calc_reg_performance(actual, predicted)
}
\description{
Internal: Calculate regression performance
}
\keyword{internal}
