% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{.sc_score_aucell}
\alias{.sc_score_aucell}
\title{Internal: AUCell scoring for single-cell}
\usage{
.sc_score_aucell(expr, genesets, ncores = 1)
}
\description{
Internal: AUCell scoring for single-cell
}
\keyword{internal}
