% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{.sc_score_mean}
\alias{.sc_score_mean}
\title{Internal: Mean expression scoring for single-cell}
\usage{
.sc_score_mean(expr, genesets)
}
\description{
Internal: Mean expression scoring for single-cell
}
\keyword{internal}
