% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{.sc_score_seurat}
\alias{.sc_score_seurat}
\title{Internal: Seurat AddModuleScore-like scoring}
\usage{
.sc_score_seurat(expr, genesets)
}
\description{
Internal: Seurat AddModuleScore-like scoring
}
\keyword{internal}
