% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{.train_ensemble}
\alias{.train_ensemble}
\title{Internal: Train Ensemble}
\usage{
.train_ensemble(X_train, y_train, X_test, y_test, task, nfolds, ...)
}
\description{
Internal: Train Ensemble
}
\keyword{internal}
