% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{.train_glmnet}
\alias{.train_glmnet}
\title{Internal: Train glmnet (Lasso/Ridge/Elastic Net)}
\usage{
.train_glmnet(X_train, y_train, X_test, y_test, task, method, nfolds, ...)
}
\description{
Internal: Train glmnet (Lasso/Ridge/Elastic Net)
}
\keyword{internal}
