% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{.train_rf}
\alias{.train_rf}
\title{Internal: Train Random Forest}
\usage{
.train_rf(X_train, y_train, X_test, y_test, task, ...)
}
\description{
Internal: Train Random Forest
}
\keyword{internal}
