% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{.train_xgboost}
\alias{.train_xgboost}
\title{Internal: Train XGBoost}
\usage{
.train_xgboost(X_train, y_train, X_test, y_test, task, nfolds, ...)
}
\description{
Internal: Train XGBoost
}
\keyword{internal}
