% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_functions.R
\name{get_all_death_genes}
\alias{get_all_death_genes}
\title{Get All Genes}
\usage{
get_all_death_genes(pathways = "all")
}
\arguments{
\item{pathways}{Character vector of pathway names. Default is "all".}
}
\value{
Character vector of unique gene symbols.
}
\description{
Get a union of all genes from specified pathways.
}
\examples{
# Get all cell death genes
all_genes <- get_all_death_genes()
length(all_genes)

# Get genes from specific pathways
genes <- get_all_death_genes(c("ferroptosis", "pyroptosis"))

}
