% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_functions.R
\name{get_pathway_info}
\alias{get_pathway_info}
\title{Get Pathway Information}
\usage{
get_pathway_info(pathway)
}
\arguments{
\item{pathway}{Character. Name of the pathway.}
}
\value{
A list containing pathway information including description,
key genes, morphology, and references.
}
\description{
Retrieve detailed information about a specific cell death pathway.
}
\examples{
get_pathway_info("ferroptosis")
get_pathway_info("cuproptosis")

}
