% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_enrichment_dot}
\alias{plot_enrichment_dot}
\title{Plot Enrichment Results - Dot Plot}
\usage{
plot_enrichment_dot(
  enrich_result,
  top_n = 10,
  size_by = "overlap",
  color_by = "pvalue"
)
}
\arguments{
\item{enrich_result}{A death_ora or death_gsea object.}

\item{top_n}{Number of top pathways to show.}

\item{size_by}{Variable for point size: "overlap" or "fold".}

\item{color_by}{Variable for point color: "pvalue" or "enrichment".}
}
\value{
A ggplot object.
}
\description{
Create a dot plot showing enrichment results.
}
