% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pathway_network}
\alias{plot_pathway_network}
\title{Plot Pathway Network}
\usage{
plot_pathway_network(scores, cor_threshold = 0.3, layout = "fr")
}
\arguments{
\item{scores}{A death_scores object.}

\item{cor_threshold}{Numeric. Minimum correlation to show edge. Default is 0.3.}

\item{layout}{Character. Network layout algorithm. Default is "fr".}
}
\value{
A ggplot object.
}
\description{
Create a network visualization of pathway relationships.
}
