% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot_time_roc}
\alias{plot_time_roc}
\title{Plot Time-dependent ROC Curves}
\usage{
plot_time_roc(roc_result, colors = NULL)
}
\arguments{
\item{roc_result}{A death_time_roc object.}

\item{colors}{Colors for different time points.}
}
\value{
A ggplot object.
}
\description{
Plot Time-dependent ROC Curves
}
