% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{print.batch_survival}
\alias{print.batch_survival}
\title{Print Method for batch_survival}
\usage{
\method{print}{batch_survival}(x, ...)
}
\arguments{
\item{x}{A batch_survival object}

\item{...}{Additional arguments}
}
\description{
Print Method for batch_survival
}
