% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{print.death_cox_multi}
\alias{print.death_cox_multi}
\title{Print Method for death_cox_multi}
\usage{
\method{print}{death_cox_multi}(x, ...)
}
\arguments{
\item{x}{A death_cox_multi object}

\item{...}{Additional arguments}
}
\description{
Print Method for death_cox_multi
}
