% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{print.death_gsea}
\alias{print.death_gsea}
\title{Print Method for death_gsea}
\usage{
\method{print}{death_gsea}(x, ...)
}
\arguments{
\item{x}{A death_gsea object}

\item{...}{Additional arguments}
}
\description{
Print Method for death_gsea
}
