% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{print.death_model}
\alias{print.death_model}
\title{Print Method for death_model}
\usage{
\method{print}{death_model}(x, ...)
}
\arguments{
\item{x}{A death_model object}

\item{...}{Additional arguments}
}
\description{
Print Method for death_model
}
