% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{print.death_ora}
\alias{print.death_ora}
\title{Print Method for death_ora}
\usage{
\method{print}{death_ora}(x, ...)
}
\arguments{
\item{x}{A death_ora object}

\item{...}{Additional arguments}
}
\description{
Print Method for death_ora
}
