% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_functions.R
\name{print.death_pathway_info}
\alias{print.death_pathway_info}
\title{Print Method for death_pathway_info}
\usage{
\method{print}{death_pathway_info}(x, ...)
}
\arguments{
\item{x}{A death_pathway_info object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for death_pathway_info
}
