% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{print.death_scores}
\alias{print.death_scores}
\title{Print Method for death_scores}
\usage{
\method{print}{death_scores}(x, ...)
}
\arguments{
\item{x}{A death_scores object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for death_scores
}
