% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{print.death_survival}
\alias{print.death_survival}
\title{Print Method for death_survival}
\usage{
\method{print}{death_survival}(x, ...)
}
\arguments{
\item{x}{A death_survival object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for death_survival
}
