% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{sc_plot_cluster_heatmap}
\alias{sc_plot_cluster_heatmap}
\title{Plot Death Scores by Cluster - Heatmap}
\usage{
sc_plot_cluster_heatmap(
  object,
  group_by = "seurat_clusters",
  pathways = "all",
  scale = TRUE
)
}
\arguments{
\item{object}{A Seurat object with death scores.}

\item{group_by}{Column in metadata to group cells by.}

\item{pathways}{Pathways to include.}

\item{scale}{Whether to scale scores. Default is TRUE.}
}
\value{
A ggplot object.
}
\description{
Create a heatmap showing average death scores for each cluster.
}
