% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{sc_plot_cluster_violin}
\alias{sc_plot_cluster_violin}
\title{Plot Death Scores by Cluster - Violin}
\usage{
sc_plot_cluster_violin(
  object,
  pathways = c("ferroptosis", "pyroptosis"),
  group_by = "seurat_clusters",
  ncol = 2
)
}
\arguments{
\item{object}{A Seurat object with death scores.}

\item{pathways}{Pathways to plot.}

\item{group_by}{Column in metadata to group cells by.}

\item{ncol}{Number of columns in plot grid.}
}
\value{
A ggplot object.
}
\description{
Create violin plots showing death score distributions for each cluster.
}
