% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{sc_plot_death_feature}
\alias{sc_plot_death_feature}
\title{Visualize Death Scores on UMAP/tSNE}
\usage{
sc_plot_death_feature(
  object,
  pathways = c("ferroptosis", "pyroptosis"),
  reduction = "umap",
  ncol = 2,
  colors = c("lightgrey", "blue", "red"),
  pt_size = 0.5
)
}
\arguments{
\item{object}{A Seurat object with death scores calculated.}

\item{pathways}{Character vector of pathways to plot.}

\item{reduction}{Which dimensionality reduction to use. Default is "umap".}

\item{ncol}{Number of columns in the plot grid.}

\item{colors}{Color gradient for scores.}

\item{pt_size}{Point size. Default is 0.5.}
}
\value{
A ggplot object or list of plots.
}
\description{
Create feature plots for death pathway scores on dimensionality reduction.
}
