% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{sc_plot_trajectory}
\alias{sc_plot_trajectory}
\title{Plot Trajectory Death Scores}
\usage{
sc_plot_trajectory(
  object,
  pseudotime_col = "pseudotime",
  pathways = c("ferroptosis", "apoptosis"),
  smooth = TRUE,
  ncol = 2
)
}
\arguments{
\item{object}{A Seurat object with pseudotime and death scores.}

\item{pseudotime_col}{Column containing pseudotime values.}

\item{pathways}{Pathways to plot.}

\item{smooth}{Whether to add smoothed trend line. Default is TRUE.}

\item{ncol}{Number of columns in plot grid.}
}
\value{
A ggplot object.
}
\description{
Visualize death scores along a pseudotime trajectory.
}
