% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_pathway}
\alias{score_pathway}
\title{Calculate Score for Single Pathway}
\usage{
score_pathway(expr, pathway, method = "zscore")
}
\arguments{
\item{expr}{Expression matrix}

\item{pathway}{Character. Name of the pathway.}

\item{method}{Scoring method. Default is "zscore".}
}
\value{
A numeric vector of scores for each sample.
}
\description{
A simplified function to calculate score for a single pathway.
}
